+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @under-control/core
 
+## 1.6.10
+
+### Patch Changes
+
+- Fix incorrect refrence for nested object returned from set-by-path
+
 ## 1.6.9
 
 ### Patch Changes